Add greedy hash approach
This version picks the congruent haplotype with the highest frequency.
Problem is, this doesn’t seem to be very accurate. I’m pushing this so
that I can revert later if needed, but I’m going to update the
algorithm to pick the matching PHASE with the highest frequency instead.